progress

+++ tools/utils.nim
@@ -42,8 +42,8 @@ proc setVKGetProc*(getProc: proc (procName: cstring): pointer {.cdecl.}) =
   vkGetProc = getProc
 
 type
-  VkHandle* = uint64
-  VkNonDispatchableHandle* = uint64
+  VkHandle* = uint
+  VkNonDispatchableHandle* = uint
   ANativeWindow = ptr object
   CAMetalLayer = ptr object
   AHardwareBuffer = ptr object
